abstract model binders to base class
