+++ src/OrchardCore/OrchardCore.Contents.Core/OrchardCore.Contents.Core.csproj
@@ -16,11 +16,12 @@
   <ItemGroup>
     <ProjectReference Include="..\OrchardCore.ContentManagement.Abstractions\OrchardCore.ContentManagement.Abstractions.csproj" />
     <ProjectReference Include="..\OrchardCore.DisplayManagement.Abstractions\OrchardCore.DisplayManagement.Abstractions.csproj" />
+    <ProjectReference Include="..\OrchardCore.Filters.Core\OrchardCore.Filters.Core.csproj" />
     <ProjectReference Include="..\OrchardCore.Infrastructure.Abstractions\OrchardCore.Infrastructure.Abstractions.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="YesSql.Filters.Query"/>
+    <PackageReference Include="YesSql.Filters.Query" />
   </ItemGroup>
 
 </Project>
