+++ src/OrchardCore/OrchardCore.Filters.Core/FilterEngineModelBinder.cs
@@ -1,0 +1,46 @@
+using System;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
+using YesSql.Filters.Query;
+
+namespace OrchardCore.Filters.Core
+{
+    public abstract class FilterEngineModelBinder<T> : IModelBinder where T : class
+    {
+        protected IQueryParser<T> _parser;
+
+        protected FilterEngineModelBinder(IQueryParser<T> parser)
+        {
+            _parser = parser;
+        }
+
+        public virtual Task BindModelAsync(ModelBindingContext bindingContext)
+        {
+            if (bindingContext == null)
+            {
+                throw new ArgumentNullException(nameof(bindingContext));
+            }
+
+            var modelName = bindingContext.ModelName;
+
+            // Try to fetch the value of the argument by name q=
+            var valueProviderResult = bindingContext.ValueProvider.GetValue(modelName);
+
+            if (valueProviderResult == ValueProviderResult.None)
+            {
+                bindingContext.Result = ModelBindingResult.Success(_parser.Parse(String.Empty));
+
+                return Task.CompletedTask;
+            }
+
+            bindingContext.ModelState.SetModelValue(modelName, valueProviderResult);
+
+            var value = valueProviderResult.FirstValue;
+
+            // When value is null or empty the parser returns an empty result.
+            bindingContext.Result = ModelBindingResult.Success(_parser.Parse(value));
+
+            return Task.CompletedTask;
+        }
+    }
+}
