+++ src/OrchardCore/OrchardCore.Users.Core/OrchardCore.Users.Core.csproj
@@ -17,13 +17,14 @@
   <ItemGroup>
     <ProjectReference Include="..\OrchardCore.Data.YesSql\OrchardCore.Data.YesSql.csproj" />
     <ProjectReference Include="..\OrchardCore.DisplayManagement.Abstractions\OrchardCore.DisplayManagement.Abstractions.csproj" />
+    <ProjectReference Include="..\OrchardCore.Filters.Core\OrchardCore.Filters.Core.csproj" />
     <ProjectReference Include="..\OrchardCore.Infrastructure.Abstractions\OrchardCore.Infrastructure.Abstractions.csproj" />
     <ProjectReference Include="..\OrchardCore.Users.Abstractions\OrchardCore.Users.Abstractions.csproj" />
   </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="YesSql.Abstractions" />
-    <PackageReference Include="YesSql.Filters.Query"/>
+    <PackageReference Include="YesSql.Filters.Query" />
   </ItemGroup>
 
 </Project>
